do two-stage validation instead of just one
first, see if the file is a well-formed CSV file per RFC4180 standard

second, normalize/transcode to utf-8 as necessary.  Normalize to "normal" CSV, as the incoming file can be a TSV/TAB file and use unconventional CSV settings.

We get the extra benefit of ensuring the file is UTF8-encoded when we use the --output option of `qsv input`